MEDICAL IMAGE SEGMENTATION AND ANALYSIS FOR DISEASE DIAGNOSIS
The healthcare industry is way more diverse than any other industries and hence deep learning and machine learning plays a significant role in this industry. Medical imaging is the process of getting images of human organs and the primary focus is to improve the reliability and the accuracy of the disease diagnosis. In this paper, the current work related to the medical image processing have been exploited to identify the effective methods for the organ segmentation as the deep learning technology is appropriate for observing the concealed patterns in images for obtaining the perfection in diagnosis.The technology not only provides real world solutions but also provides good accuracy in the process of medical image analysis and the diagnosis system. Prediction of the disease of the liver tumor have been performed to show the predictability and the accuracy of liver tumor to compare the performance metrics by the evaluation.